refactor shared logic into ExtendedBaseRequest class
