+++ src/main/java/com/auth0/net/FormDataRequest.java
@@ -4,8 +4,6 @@ import java.io.File;
 
 interface FormDataRequest<T> extends Request<T> {
 
-    FormDataRequest<T> addHeader(String name, String value);
-
     FormDataRequest<T> addPart(String name, String value);
 
     FormDataRequest<T> addPart(String name, File file, String dataType);
