+++ src/main/java/com/auth0/net/MultipartRequest.java
@@ -1,45 +1,30 @@
 package com.auth0.net;
 
-import com.auth0.exception.APIException;
-import com.auth0.exception.Auth0Exception;
-import com.auth0.exception.RateLimitException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.type.MapType;
-import okhttp3.Request;
 import okhttp3.*;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
 
 import static com.auth0.utils.Asserts.assertNotNull;
 
 @SuppressWarnings("WeakerAccess")
-public class MultipartRequest<T> extends BaseRequest<T> implements FormDataRequest<T> {
+public class MultipartRequest<T> extends ExtendedBaseRequest<T> implements FormDataRequest<T> {
 
     private static final String CONTENT_TYPE_FORM_DATA = "multipart/form-data";
+
     private final MultipartBody.Builder bodyBuilder;
-
-    private final String url;
-    private final String method;
-    private final HashMap<String, String> headers;
-    private final ObjectMapper mapper;
     private final TypeReference<T> tType;
-
-    private static final int STATUS_CODE_TOO_MANY_REQUEST = 429;
+    private final ObjectMapper mapper;
 
     MultipartRequest(OkHttpClient client, String url, String method, ObjectMapper mapper, TypeReference<T> tType, MultipartBody.Builder multipartBuilder) {
-        super(client);
+        super(client, url, method, mapper);
         if ("GET".equalsIgnoreCase(method)) {
             throw new IllegalArgumentException("The HTTP method GET is not supported");
         }
-        this.url = url;
-        this.method = method;
         this.mapper = mapper;
         this.tType = tType;
-        this.headers = new HashMap<>();
         this.bodyBuilder = multipartBuilder
                 .setType(MultipartBody.FORM);
     }
@@ -49,8 +34,25 @@ public class MultipartRequest<T> extends BaseRequest<T> implements FormDataReque
     }
 
     @Override
-    public FormDataRequest<T> addHeader(String name, String value) {
-        headers.put(name, value);
+    protected String getContentType() {
+        return CONTENT_TYPE_FORM_DATA;
+    }
+
+    @Override
+    protected RequestBody createRequestBody() throws IOException {
+        return bodyBuilder.build();
+    }
+
+    @Override
+    protected T readResponseBody(ResponseBody body) throws IOException {
+        String payload = body.string();
+        return mapper.readValue(payload, tType);
+    }
+
+    @Override
+    public MultipartRequest<T> addHeader(String name, String value) {
+        //This is to avoid returning a different type
+        super.addHeader(name, value);
         return this;
     }
 
@@ -70,67 +72,5 @@ public class MultipartRequest<T> extends BaseRequest<T> implements FormDataReque
         assertNotNull(value, "value");
         bodyBuilder.addFormDataPart(name, value);
         return this;
-    }
-
-    @Override
-    protected Request createRequest() throws Auth0Exception {
-        //FIXME: Copies from #com.auth0.net.CustomRequest
-        MultipartBody body;
-        try {
-            body = bodyBuilder.build();
-        } catch (IllegalStateException e) {
-            throw new Auth0Exception("Couldn't create the request body.", e);
-        }
-        Request.Builder builder = new Request.Builder()
-                .url(url)
-                .method(method, body);
-        //TODO: Move to "getBody" abstract method. Handle parts size.
-        for (Map.Entry<String, String> e : headers.entrySet()) {
-            builder.addHeader(e.getKey(), e.getValue());
-        }
-        builder.addHeader("Content-Type", CONTENT_TYPE_FORM_DATA);
-        return builder.build();
-    }
-
-    @Override
-    protected T parseResponse(Response response) throws Auth0Exception {
-        //FIXME: Copies from #com.auth0.net.CustomRequest
-        if (!response.isSuccessful()) {
-            throw createResponseException(response);
-        }
-
-        try (ResponseBody body = response.body()) {
-            String payload = body.string();
-            //TODO: Move to "readBody" abstract method.
-            return mapper.readValue(payload, tType);
-        } catch (IOException e) {
-            throw new APIException("Failed to parse json body", response.code(), e);
-        }
-    }
-
-    protected Auth0Exception createResponseException(Response response) {
-        //FIXME: Copies from #com.auth0.net.CustomRequest
-        if (response.code() == STATUS_CODE_TOO_MANY_REQUEST) {
-            return createRateLimitException(response);
-        }
-
-        String payload = null;
-        try (ResponseBody body = response.body()) {
-            payload = body.string();
-            MapType mapType = mapper.getTypeFactory().constructMapType(HashMap.class, String.class, Object.class);
-            Map<String, Object> values = mapper.readValue(payload, mapType);
-            return new APIException(values, response.code());
-        } catch (IOException e) {
-            return new APIException(payload, response.code(), e);
-        }
-    }
-
-    private RateLimitException createRateLimitException(Response response) {
-        //FIXME: Copies from #com.auth0.net.CustomRequest
-        // -1 as default value if the header could not be found.
-        long limit = Long.parseLong(response.header("X-RateLimit-Limit", "-1"));
-        long remaining = Long.parseLong(response.header("X-RateLimit-Remaining", "-1"));
-        long reset = Long.parseLong(response.header("X-RateLimit-Reset", "-1"));
-        return new RateLimitException(limit, remaining, reset);
     }
 }
